+++ arduino-gps-mappy/host-server/GPSHandler.py
@@ -6,6 +6,7 @@ class GPSHandler:
     def __init__(self):
         self.lat = 0
         self.lng = 0
+        self.upt = 0
         self.isValid = False
 
     def readPacket(self, packet):
@@ -17,14 +18,16 @@ class GPSHandler:
             print("GPS Handler: readPacket: invalid packet type")
             self.isValid = False
             return
-        if len(packet[1]) != 4:
+        if len(packet[1]) != 6:
             print("GPS Handler: readPacket: invalid packet data size")
             self.isValid = False
             return
 
         latitudeBytes = bytes(packet[1][0:2])
-        longitudeBytes = bytes(packet[1][2:])
+        longitudeBytes = bytes(packet[1][2:4])
+        updateBytes = bytes(packet[1][4:])
 
         self.lat = float(struct.unpack("<f", latitudeBytes)[0])
         self.lng = float(struct.unpack("<f", longitudeBytes)[0])
+        self.upt = int(struct.unpack("<I", updateBytes)[0])
         self.isValid = True