+++ arduino-gps-mappy/host-server/main.py
@@ -6,14 +6,6 @@ from PolygonHandler import PolygonHandler
 from interface import makeApp
 
 
-def serialInterface(serialHandler):
-    try:
-        print()
-    except serial.SerialException as e:
-        print(f"Error: {e}")
-        return
-
-
 port = '/dev/ttyUSB0'  # 'dev/ttyUSB0' on linux
 if platform.system() == "Linux":
     port = '/dev/ttyUSB0'
@@ -21,12 +13,24 @@ elif platform.system() == "Windows":
     port = 'COM3'
 
 baudRate = 19200
-serialPort = SerialHandler(baudRate=baudRate, port=port)
-polygonHandler = PolygonHandler()
 
-app = makeApp()
+
+def serialInterface():
+    global port
+    global baudRate
+    global serialPort
+    global polygonHandler
+    try:
+        serialPort = SerialHandler(baudRate=baudRate, port=port)
+        polygonHandler = PolygonHandler()
+        print()
+    except serial.SerialException as e:
+        print(f"Error: {e}")
+        return
+
 
 running = True
 serialThread = threading.Thread(target=serialInterface, daemon=True)
 serialThread.start()
+app = makeApp()
 app.run()