+++ CHANGELOG.md
@@ -1,2 +1,2 @@
 # Mappy
-- host server got gps parser, serial handler, polygon parser modules
+- host server got serial time out, can handle partialy sent message, got polygon CLI partially implemented